Added files to this upload
Added files to this upload to deploy a blog that I have been having in development for the past year. I would like to continue developing this blog and further push my limits as a web developer. It will also help me to learn Markdown and become more proficient in the language.